Remove abstract, change weight to double,

+++ src/main/java/workouts/GymSet.java
@@ -24,7 +24,7 @@ public class GymSet {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
